+++ changelog.md
@@ -7,6 +7,7 @@ Actual version: "0.0.0"
 
 ## 0.0
 
+* [added] Pixel size estimation based on incision size
 * [changed] Video has lower resolution to be loaded faster
 * [changed] Video and graph are concatenated horizontaly
 * [added] Poster to see first frame of the video before the video is loaded